+++ tests/packages/test_dependency.py
@@ -66,5 +66,30 @@ def test_to_pep_508():
 
     result = dependency.to_pep_508()
     assert result == 'Django (>=1.23.0.0,<2.0.0.0); ' \
-                     '(python_version>="2.7.0.0" and python_version<"2.8.0.0") ' \
-                     'or (python_version>="3.6.0.0" and python_version<"4.0.0.0")'
+                     '(python_version >= "2.7.0.0" and python_version < "2.8.0.0") ' \
+                     'or (python_version >= "3.6.0.0" and python_version < "4.0.0.0")'
+
+
+def test_to_pep_508_in_extras():
+    dependency = Dependency('Django', '^1.23')
+    dependency.in_extras.append('foo')
+
+    result = dependency.to_pep_508()
+    assert result == 'Django (>=1.23.0.0,<2.0.0.0); extra == "foo"'
+
+    dependency.in_extras.append('bar')
+
+    result = dependency.to_pep_508()
+    assert result == 'Django (>=1.23.0.0,<2.0.0.0); extra == "foo" or extra == "bar"'
+
+    dependency.python_versions = '~2.7 || ^3.6'
+
+    result = dependency.to_pep_508()
+    assert result == (
+        'Django (>=1.23.0.0,<2.0.0.0); '
+        '('
+        '(python_version >= "2.7.0.0" and python_version < "2.8.0.0") '
+        'or (python_version >= "3.6.0.0" and python_version < "4.0.0.0")'
+        ') '
+        'and (extra == "foo" or extra == "bar")'
+    )